make redisdb a fixture factory
