+++ pytest_dbfixtures/factories.py
@@ -1,7 +1,7 @@
 import pytest
 from summon_process.executors import TCPCoordinatedExecutor
 
-from pytest_dbfixtures.utils import get_config
+from pytest_dbfixtures.utils import get_config, try_import
 
 
 def redis_proc(executable=None, params=None, config_file=None,
@@ -47,3 +47,29 @@ def redis_proc(executable=None, params=None, config_file=None,
         return redis_executor
 
     return redis_proc_fixture
+
+
+def redisdb(process_fixture_name, host=None, port=None, db=None):
+
+    @pytest.fixture
+    @pytest.mark.usefixtures(process_fixture_name)
+    def redisdb_factory(request, redis_proc):
+        redis, config = try_import('redis', request)
+
+        redis_host = host
+        if not redis_host:
+            redis_host = config.redis.host
+
+        redis_port = port
+        if not redis_port:
+            redis_port = config.redis.port
+
+        redis_db = db
+        if not redis_db:
+            redis_db = config.redis.db
+
+        redis_client = redis.Redis(redis_host, redis_port, redis_db)
+        request.addfinalizer(redis_client.flushall)
+        return redis_client
+
+    return redisdb_factory
