+++ pytest_dbfixtures/utils.py
@@ -1,8 +1,23 @@
-from pymlconf import ConfigManager
+import importlib
 
-__all__ = ['get_config']
+from pymlconf import ConfigManager
 
 
 def get_config(request):
     config_name = request.config.getvalue('db_conf')
     return ConfigManager(files=[config_name])
+
+
+def try_import(module, request, pypi_package=None):
+    try:
+        i = importlib.import_module(module)
+    except ImportError:
+        raise ImportError(
+            'Please install {0} package.\n'
+            'pip install -U {0}'.format(
+                pypi_package or module
+            )
+        )
+    else:
+
+        return i, get_config(request)
